feat: onMessage event handler
send message embed to channel and user
