+++ src/events/onMessage.js
@@ -1,0 +1,6 @@
+const { checkRestricted } = require('../modules/restrictedWords');
+
+const onMessage = async (message) => {
+	await checkRestricted(message);
+};
+module.exports = { onMessage };
